feature: move command to beta package
